requirements.txt
streamlit>=1.35,<1.40
pandas>=2.0,<2.3
requests>=2.31,<2.33
python-dateutil>=2.8.2,<2.9
duckduckgo_search>=5.3.0,<5.4
